move tests to correct directories
