+++ src/applications/gi/tests/components/search/SearchResult.unit.spec.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { expect } from 'chai';
 import { shallow } from 'enzyme';
 
-import { SearchResult } from '../../components/search/SearchResult';
+import { SearchResult } from '../../../components/search/SearchResult';
 
 const result = {
   name: 'BENNETT COLLEGE',
